Exceptions fix

+++ TPE/Userland/SampleCodeModule/commands.c
@@ -27,9 +27,8 @@ void clearScreen(){
 
 
 void time(){
-	int currTime[6];
+	uint64_t currTime[6];
 	sys_time(currTime);
-	// uint64_t * currTime = os_time();		returns an array with the current hour, minutes, and seconds
 
 	char time[9];					//pass the array to a string to print
 	time[0] = currTime[0] + '0';
